Fixed compiler warnings.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@1998 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/intl.cpp
@@ -1660,7 +1660,7 @@ unsigned short TextTypeNC::mbtowc(WCHAR *wc, unsigned char *ptr, unsigned short 
 	}
 	if (wc)
 		*wc = 0;
-	return -1;					/* No more characters */
+	return (unsigned short)-1;			/* No more characters */
 }
 
 unsigned short TextTypeMB::contains(TDBB a, unsigned char *b,
@@ -1731,7 +1731,7 @@ unsigned short TextTypeMB::mbtowc(WCHAR *wc, unsigned char *ptr, unsigned short 
 	}
 	if (wc)
 		*wc = 0;
-	return -1;					/* No more characters */
+	return (unsigned short)-1;			/* No more characters */
 }
 
 unsigned short TextTypeWC::to_wc(unsigned char *pDestUC,
@@ -1844,7 +1844,7 @@ unsigned short TextTypeWC::mbtowc(WCHAR *wc, unsigned char *ptr, unsigned short 
 	}
 	if (wc)
 		*wc = 0;
-	return -1;					/* No more characters */
+	return (unsigned short)-1;			/* No more characters */
 }
